crates/sel4-initialize-tls: Create set_thread_pointer module
